# docs/openapi.json
{
  "components": {
    "schemas": {
      "ArtifactOut": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Размер файла; null, если его нет",
            "title": "Size Bytes"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "name",
          "url"
        ],
        "title": "ArtifactOut",
        "type": "object"
      },
      "Body_create_run_api_v1_runs_post": {
        "properties": {
          "extra": {
            "anyOf": [
              {
                "items": {
                  "contentMediaType": "application/octet-stream",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Остальные чертежи комплекта (DXF или DWG): склеиваются с основным",
            "title": "Extra"
          },
          "file": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Чертёж DXF или DWG; или street, но не оба",
            "title": "File"
          },
          "inventory": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Перечётная ведомость .xls или .xlsx: существующие деревья в квотах",
            "title": "Inventory"
          },
          "overrides": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "JSON-объект параметров поверх профиля",
            "title": "Overrides"
          },
          "profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Профиль параметров из /meta",
            "title": "Profile"
          },
          "street": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Улица пилотного проекта из /streets (slug); или file",
            "title": "Street"
          }
        },
        "title": "Body_create_run_api_v1_runs_post",
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "ProgressOut": {
        "description": "Ход прогона: есть только пока он идёт.\n\nДоля и остаток - оценка по весам этапов и прошедшему времени; подоснова для карты\n(`basemap.geojson`) появляется в `artifacts` раньше остальных файлов.",
        "properties": {
          "elapsed_s": {
            "description": "Секунд с начала расчёта",
            "title": "Elapsed S",
            "type": "number"
          },
          "eta_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Оценка остатка в секундах",
            "title": "Eta S"
          },
          "fraction": {
            "description": "Доля сделанного, 0..1",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Fraction",
            "type": "number"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Текущий этап или null до первого этапа",
            "title": "Stage"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepOut"
            },
            "title": "Steps",
            "type": "array"
          },
          "title": {
            "description": "Название этапа для человека",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "stage",
          "title",
          "fraction",
          "elapsed_s",
          "eta_s",
          "steps"
        ],
        "title": "ProgressOut",
        "type": "object"
      },
...
      "RunListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "RunListOut",
        "type": "object"
      },
      "RunOut": {
        "properties": {
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactOut"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "overrides": {
            "additionalProperties": true,
            "title": "Overrides",
            "type": "object"
          },
          "profile": {
            "title": "Profile",
            "type": "string"
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProgressOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "Ход расчёта; null, пока прогон в очереди или уже закончен"
          },
          "source_name": {
            "title": "Source Name",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/RunState"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "state",
          "source_name",
          "profile",
          "overrides",
          "created_at",
          "updated_at"
        ],
        "title": "RunOut",
        "type": "object"
      },
      "RunState": {
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed"
        ],
        "title": "RunState",
        "type": "string"
      },
...
      "StepOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Длительность завершённого этапа",
            "title": "Ms"
          },
          "state": {
            "enum": [
              "done",
              "active",
              "pending"
            ],
            "title": "State",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "state"
        ],
        "title": "StepOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Прогон принимает DXF или DWG, строит план посадок на отдельных слоях GREEN_* и возвращает DXF, объяснения со ссылками на пункты НПА и отчёт о целостности исходных слоёв.",
    "summary": "Автопроектирование озеленения улиц с учётом подземных сетей и норм.",
    "title": "green API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/runs": {
      "get": {
        "description": "Последние прогоны, новые первыми.",
        "operationId": "list_runs_api_v1_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "description": "Принять чертёж или улицу из каталога и поставить прогон в очередь.\n\nИсточник ровно один: свой чертёж (`file`, к нему комплект `extra`) или улица пилотного\nпроекта (`street`, её комплект уже лежит в каталоге). Статус: GET /runs/{id}.",
        "operationId": "create_run_api_v1_runs_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_run_api_v1_runs_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Create Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/demo": {
      "post": {
        "description": "Прогон встроенного фрагмента улицы Берзарина с профилем по умолчанию.\n\nФрагмент лежит в пакете: на стенде жюри датасета нет, а показывать сервис надо с\nпервого клика, не заставляя искать DXF. Статус: GET /runs/{id}.",
        "operationId": "create_demo_run_api_v1_runs_demo_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Create Demo Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "description": "Статус прогона, сводка и ссылки на артефакты.",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/artifacts/{name}": {
      "get": {
        "description": "Скачать артефакт: result.dxf, plan.json, interpretations.csv и другие.\n\nТекстовые артефакты отдаются обычным ответом, а не FileResponse, намеренно. Granian\nумеет отправлять файл в обход ASGI-конвейера (расширение pathsend), и тогда сжатие\nmiddleware не применяется: подоснова Берзарина уехала бы в браузер на 12,9 МБ вместо\n1,3 МБ. Крупный result.dxf, наоборот, скачивают целиком и жать его незачем.",
        "operationId": "get_artifact_api_v1_runs__run_id__artifacts__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Artifact",
        "tags": [
          "runs"
        ]
      }
    },
...
  },
...
}